refactor(ui): ✨ refactor search
